Feat Reset_password_view | Commit 27
