+++ knownProgrammer/accounts/static/css/popup.css
@@ -79,4 +79,3 @@
     line-height: 30px;
     text-decoration: none;
 }
-
