+++ knownProgrammer/accounts/templates/accounts/login.html
@@ -28,8 +28,8 @@
             {% endif %}
           <button type="submit" class="btn btn-secondary">Submit</button>
         </form>
+        <br>
+        <a href="{% url 'reset_password' %}" style="color: blue; font-size: 14px; text-decoration: none">Forget your password?</a>
     </div>
     {% endif %}
-
-
 {% endblock body %}
