+++ knownProgrammer/accounts/templates/accounts/password_reset_complete.html
@@ -1,0 +1,19 @@
+{% extends "accounts/base.html" %}
+{% load static %}
+
+{% block title %}Password Reset Complete{% endblock title %}
+
+{% block body %}
+<div class="container mt-5">
+    <div class="row justify-content-center">
+        <div class="col-md-6">
+            <div class="card">
+                <div class="card-header" style="background-color: #696969; color: #fff;">Password Reset Complete</div>
+                <div class="card-body" style="padding: 20px;">
+                    <p class="card-text">Your password has been successfully reset. You can now <a href="{% url 'login' %}">log in</a> with your new password.</p>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock body %}
