+++ knownProgrammer/accounts/templates/accounts/password_reset_confirm.html
@@ -1,0 +1,35 @@
+{% extends "accounts/base.html" %}
+{% load static %}
+
+{% block title %}Confirm Password Reset{% endblock title %}
+
+{% block body %}
+<div class="container mt-5">
+    <div class="row justify-content-center">
+        <div class="col-md-6">
+            <div class="card">
+                <style>
+                    .btn-warning {
+                        background-color: #696969;
+                        border: none;
+                        color: #ffffff;
+                    }
+                    .btn-warning:hover {
+                    background-color: #3D3D3D;
+                    color: #ffffff;
+                    }
+                </style>
+                <div class="card-header" style="background-color: #696969; color: #ffffff;">Confirm Password Reset</div>
+                <div class="card-body" style="padding: 20px;">
+                    <p class="card-text">Enter your new password twice to confirm.</p>
+                    <form class="form" method="POST">
+                        {% csrf_token %}
+                        {{ form.as_p }}
+                        <input type="submit" class="btn btn-warning" value="Confirm">
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock body %}
