+++ knownProgrammer/accounts/templates/accounts/password_reset_done.html
@@ -1,0 +1,19 @@
+{% extends "accounts/base.html" %}
+{% load static %}
+
+{% block title %}Password Reset Done{% endblock title %}
+
+{% block body %}
+<div class="container mt-5">
+    <div class="row justify-content-center">
+        <div class="col-md-6">
+            <div class="card">
+                <div class="card-header" style="background-color: #696969; color: #fff;">Password Reset Email Sent</div>
+                <div class="card-body" style="padding: 20px;">
+                    <p class="card-text">We've emailed you instructions for setting a new password. Check your inbox and follow the instructions to reset your password.</p>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock body %}
