+++ knownProgrammer/accounts/templates/accounts/reset_password.html
@@ -1,0 +1,66 @@
+{% extends "accounts/base.html" %}
+{% load static %}
+
+{% block title %}Password Reset{% endblock title %}
+
+{% block body %}
+<div class="container mt-5">
+    <div class="row justify-content-center">
+        <div class="col-md-6">
+            <style>
+                .card {
+                    border: 1px solid #ccc;
+                    border-radius: 10px;
+                    margin-top: 20px;
+                }
+
+                .card-header {
+                    background-color: #696969;
+                    color: #fff;
+                    padding: 10px 15px;
+                    border-top-left-radius: 10px;
+                    border-top-right-radius: 10px;
+                }
+
+                .card-body {
+                    padding: 20px;
+                }
+
+                .card-text {
+                    margin-bottom: 15px;
+                }
+
+                .my-button {
+                    display: inline-block;
+                    padding: 8px 16px;
+                    font-size: 14px;
+                    text-align: center;
+                    text-decoration: none;
+                    border: none;
+                    border-radius: 4px;
+                    background-color: #696969;
+                    color: #ffffff;
+                }
+
+                .my-button:hover,
+                .my-button:focus {
+                    background-color: #3D3D3D;
+                    color: #ffffff;
+                }
+
+            </style>
+            <div class="card">
+                <div class="card-header">Password Reset</div>
+                <div class="card-body">
+                    <p class="card-text">Forgotten your password? Enter your email address below, and we'll email instructions for setting a new one.</p>
+                    <form class="form" method="POST">
+                        {% csrf_token %}
+                        {{ form.as_p }}
+                        <input type="submit" class="my-button" value="Reset Password">
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock body %}
