+++ knownProgrammer/envtemplate.txt
@@ -1,6 +1,18 @@
 SECRET_KEY = ""
-DEBUG = True
-DB_USER = ""
-DB_PASSWORD = ""
+DEBUG =
+
+DB_USER =
+DB_PASSWORD =
 DB_PORT =
-DB_HOST = ""
+DB_HOST =
+
+CLOUD_NAME =
+API_KEY =
+API_SECRET =
+
+EMAIL_BACKEND = ''
+EMAIL_HOST = ''
+EMAIL_PORT =
+EMAIL_USE_TLS =
+EMAIL_HOST_USER = ''
+EMAIL_HOST_PASSWORD = ''
